assistant-implementation/requirements.txt
# FastAPI Nova Sonic Assistant Requirements
# Production dependencies for ECS Fargate

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.20

# AWS SDK - MANDATORY BIDIRECTIONAL STREAMING
boto3==1.36.7
botocore==1.36.7
aws_sdk_bedrock_runtime==0.0.2
smithy-aws-core==0.0.3
smithy-aws-event-stream==0.0.1
smithy-core==0.0.2
smithy-http==0.0.1
smithy-json==0.0.1
aws-sdk-signers==0.0.3
awscrt>=0.27.6
ijson>=3.4.0

# Redis for session management
redis[hiredis]==5.2.1

# Audio processing
numpy==1.26.4
scipy==1.14.1
librosa==0.10.2.post1
soundfile==0.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
python-json-logger==3.2.0
prometheus-client==0.21.0

# Performance
aiocache==0.12.3
orjson==3.10.12

# Development/Testing
pytest==8.3.5
pytest-asyncio==0.25.2
httpx==0.28.1